auth_service/requirements.txt
Flask==2.3.3
pymongo==4.6.0
Flask-JWT-Extended==4.5.3
bcrypt==4.1.1
python-consul==1.1.0

